fix(gcdb-updater): read HTTP status into a long, not an int
curl_easy_getinfo(CURLINFO_RESPONSE_CODE) writes a long through the
pointer it is given. Handing it an int lets it write 8 bytes into 4 on
LP64, and the 4 bytes past the variable belong to whatever the compiler
parked next to it on the stack.

Under clang that neighbour is the spilled `ctx` argument, so the status
line "HTTP/2 200 " zeroes its low half - 0x5555555xxxxx becomes
0x555500000000 - and the next store through ctx segfaults, on the first
header of every request. gcc happens to lay the frame out so the write
lands somewhere harmless, which is why this only ever showed up on one
compiler.

Verified against the crashing build: test_gcdb_update went from a
reproducible SIGSEGV to passing.

+++ device/gcdb-updater/gamecontrollerdb_updater.c
@@ -133,12 +133,12 @@ static void write_header_lines(task_context *ctx) {
 
 static size_t header_cb(char *buffer, size_t size, size_t nitems, task_context *ctx) {
     if (ctx->status == 0) {
-        int status = 0;
+        long status = 0;
         curl_easy_getinfo(ctx->curl, CURLINFO_RESPONSE_CODE, &status);
         if (status == 301 || status == 302) {
             return size * nitems;
         }
-        ctx->status = status;
+        ctx->status = (int) status;
         if (status != 200) {
             return 0;
         }
